+++ src/StateSmith/Output/Algos/Balanced1/AlgoBalanced1Settings.cs
@@ -44,13 +44,19 @@ public class AlgoBalanced1Settings
     internal bool outputSwitchDefault = false;
 
     /// <summary>
-    /// Set to false if don't want the event ID to string function to be generated and output.
+    /// Set to false if you don't want the event ID to string function to be generated and output.
     /// https://github.com/StateSmith/StateSmith/issues/181
     /// </summary>
     public bool outputEventIdToStringFunction = true;
 
     /// <summary>
-    /// Set to false if don't want the state ID to string function to be generated and output.
+    /// Set to false if you don't want the "is event ID valid" function to be generated and output.
+    /// https://github.com/StateSmith/StateSmith/issues/473
+    /// </summary>
+    public bool outputEventIdIsValidFunction = true;
+
+    /// <summary>
+    /// Set to false if you don't want the state ID to string function to be generated and output.
     /// https://github.com/StateSmith/StateSmith/issues/181
     /// </summary>
     public bool outputStateIdToStringFunction = true;