+++ src/StateSmith/Output/Algos/Balanced1/IAlgoEventIdFuncGenerator.cs
@@ -4,7 +4,8 @@ using StateSmith.SmGraph;
 
 namespace StateSmith.Output.Algos.Balanced1;
 
-public interface IAlgoEventIdToString
+public interface IAlgoEventIdFuncGenerator
 {
     void CreateEventIdToStringFunction(OutputFile file, StateMachine sm);
+    void CreateEventIdIsValidFunction(OutputFile file, StateMachine sm);
 }